feat: FW-1060 post translation
https://mask.atlassian.net/browse/FW-1060
